SDE MultiVersionedViews

+++ gView.DataSources.OSGeo/DataSources/GDAL/RasterClass.cs
@@ -416,7 +416,7 @@ namespace gView.DataSources.GDAL
                                 break;
                         }
                         band.ReadRaster(x, y, wWidth, wHeight,
-                            new IntPtr(buf.ToInt32() + ch),
+                            new IntPtr(buf.ToInt64() + ch),
                             iWidth, iHeight, OSGeo.GDAL.DataType.GDT_Byte, pixelSpace, stride);
 
                         band.Dispose();
@@ -497,7 +497,7 @@ namespace gView.DataSources.GDAL
                                 break;
                         }
                         band.ReadRaster(x, y, wWidth, wHeight,
-                                new IntPtr(buf.ToInt32() + ch),
+                                new IntPtr(buf.ToInt64() + ch),
                                 iWidth, iHeight, OSGeo.GDAL.DataType.GDT_Byte, pixelSpace, stride);
 
                         band.Dispose();
